requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
apipkg==1.5
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
coverage==5.5
crayons==0.4.0
cryptography==3.4.7
execnet==1.8.0
idna==2.10
iniconfig==1.1.1
makefun==1.11.3
more-itertools==8.7.0
packaging==20.9
pdfminer==20191125
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pyperclip==1.8.2
pypsrp==0.5.0
pyspnego==0.1.5
pytest==6.2.3
pytest-cov==2.11.1
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-rerunfailures==9.1.1
pytest-steps==1.7.3
pytest-xdist==2.2.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
tblib==1.7.0
toml==0.10.2
urllib3==1.26.4
webdriver-manager==3.4.1
win32-setctime==1.0.3
wrapt==1.12.1
bs4~=0.0.1
beautifulsoup4~=4.9.3